working on the map
I'm working on the map layout, I need to look at a random area to set
the test zone.
I was able to figure out how to click a button and get it to draw, and
so on.

There needs to be a button which centers the map back to the starting
point.